messaging_app/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
